+++ packages/tdl-tdlib-addon/td.cpp
@@ -12,10 +12,6 @@ typedef const char * (*td_json_client_receive_t)(void *client, double timeout);
 typedef const char * (*td_json_client_execute_t)(void *client, const char *request);
 typedef void (*td_json_client_destroy_t)(void *client);
 
-typedef int (*td_set_log_file_path_t)(const char *file_path);
-typedef void (*td_set_log_max_file_size_t)(long long max_file_size);
-typedef void (*td_set_log_verbosity_level_t)(int new_verbosity_level);
-
 typedef void (*td_log_fatal_error_callback_ptr)(const char *error_message);
 typedef void (*td_set_log_fatal_error_callback_t)(td_log_fatal_error_callback_ptr callback);
 
@@ -25,9 +21,6 @@ td_json_client_receive_t td_json_client_receive;
 td_json_client_execute_t td_json_client_execute;
 td_json_client_destroy_t td_json_client_destroy;
 
-td_set_log_file_path_t td_set_log_file_path;
-td_set_log_max_file_size_t td_set_log_max_file_size;
-td_set_log_verbosity_level_t td_set_log_verbosity_level;
 td_set_log_fatal_error_callback_t td_set_log_fatal_error_callback;
 
 Napi::ArrayBuffer td_client_create(const Napi::CallbackInfo& info) {
@@ -102,24 +95,6 @@ void td_client_destroy(const Napi::CallbackInfo& info) {
   td_json_client_destroy(client);
 }
 
-Napi::Number td_set_file_path(const Napi::CallbackInfo& info) {
-  Napi::Env env = info.Env();
-  std::string filepath = info[0].As<Napi::String>().Utf8Value();
-  int out = td_set_log_file_path(filepath.c_str());
-  return Napi::Number::New(env, out);
-}
-
-void td_set_max_file_size(const Napi::CallbackInfo& info) {
-  std::string max_file_size_str = info[0].As<Napi::String>().Utf8Value();
-  long long max_file_size = stoll(max_file_size_str);
-  td_set_log_max_file_size(max_file_size);
-}
-
-void td_set_verbosity_level(const Napi::CallbackInfo& info) {
-  auto level = (int) info[0].As<Napi::Number>().Int32Value();
-  td_set_log_verbosity_level(level);
-}
-
 Napi::FunctionReference js_fatal_cb;
 
 // This should work, but I haven't tested it
@@ -166,9 +141,6 @@ void load_tdjson(const Napi::CallbackInfo& info) {
   FINDFUNC(td_json_client_receive);
   FINDFUNC(td_json_client_execute);
   FINDFUNC(td_json_client_destroy);
-  FINDFUNC(td_set_log_file_path);
-  FINDFUNC(td_set_log_max_file_size);
-  FINDFUNC(td_set_log_verbosity_level);
   FINDFUNC(td_set_log_fatal_error_callback);
 }
 
@@ -183,12 +155,6 @@ Napi::Object Init(Napi::Env env, Napi::Object exports) {
     "td_client_execute", Napi::Function::New(env, td_client_execute));
   exports.Set(
     "td_client_destroy", Napi::Function::New(env, td_client_destroy));
-  exports.Set(
-    "td_set_file_path", Napi::Function::New(env, td_set_file_path));
-  exports.Set(
-    "td_set_max_file_size", Napi::Function::New(env, td_set_max_file_size));
-  exports.Set(
-    "td_set_verbosity_level", Napi::Function::New(env, td_set_verbosity_level));
   exports.Set(
     "td_set_fatal_error_callback", Napi::Function::New(env, td_set_fatal_error_callback));
   exports.Set(