+++ packages/tdl-tdlib-ffi/src/index.js
@@ -10,7 +10,7 @@ const debug = Debug('tdl-tdlib-ffi')
 
 const buildQuery = (query: Object) => {
   const buffer = Buffer.from(JSON.stringify(query) + '\0', 'utf-8')
-  // $FlowOff
+  // $FlowIgnore[prop-missing]
   buffer.type = ref.types.CString
   return buffer
 }
@@ -40,7 +40,11 @@ export class TDLib implements ITDLibJSON {
       })
   }
 
-  async create (): Promise<TDLibClient> {
+  getName (): string {
+    return 'ffi'
+  }
+
+  create (): TDLibClient {
     debug('create')
     return this._tdlib.td_json_client_create()
   }
@@ -68,21 +72,6 @@ export class TDLib implements ITDLibJSON {
   destroy (client: TDLibClient): void {
     debug('destroy')
     this._tdlib.td_json_client_destroy(client)
-  }
-
-  setLogFilePath (path: string): number {
-    debug('setLogFilePath', path)
-    return this._tdlib.td_set_log_file_path(path)
-  }
-
-  setLogMaxFileSize (maxFileSize: number | string): void {
-    debug('setLogMaxFileSize', maxFileSize)
-    this._tdlib.td_set_log_max_file_size(maxFileSize)
-  }
-
-  setLogVerbosityLevel (verbosity: number): void {
-    debug('setLogVerbosityLevel', verbosity)
-    this._tdlib.td_set_log_verbosity_level(verbosity)
   }
 
   setLogFatalErrorCallback (fn: null | (errorMessage: string) => void): void {